+++ src/app/@modal/(.)auth/login/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { AuthModal } from "@/components/global/auth/client";
+
+export default function Page() {
+  return <AuthModal />;
+}
