+++ src/components/global/auth/server/index.tsx
@@ -1,12 +1,10 @@
 "server only";
 
-import { Mail } from "lucide-react";
+import { LogOutIcon, Mail } from "lucide-react";
+import Link from "next/link";
 import { Session } from "next-auth";
 import React from "react";
-import {
-  AuthButton,
-  SignOutButton,
-} from "@/components/global/parts/auth-button";
+
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
 import {
@@ -14,9 +12,28 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from "@/components/ui/popover";
+import { signOutOauth } from "@/lib/auth/actions";
 
 export function UnAuthorized() {
-  return <AuthButton />;
+  return (
+    <Button asChild size="sm">
+      <Link href="/auth/login">LogIn</Link>
+    </Button>
+  );
+}
+
+export function SignOutButton() {
+  return (
+    <form action={signOutOauth}>
+      <button
+        className="flex w-full items-center gap-x-2 text-sm"
+        type="submit"
+      >
+        <LogOutIcon className="inline-block h-4 w-4" />
+        <span>ログアウト</span>
+      </button>
+    </form>
+  );
 }
 
 export function Authorized({ session }: { session: Session }) {
